refactor: reorganize neural_networks into siamese/scripts subpackages
Introduce a `siamese` subpackage under `pyvisim.neural_networks` holding the
`SiameseNeuralNetwork` model and a new `backbones.py` with the `ResNetBackbone`
class (moved verbatim out of the training script). The training entry point is
relocated to a `scripts` subpackage.

`SiameseNeuralNetwork` now defaults its `backbone` to a pretrained
`ResNetBackbone` when `None` is passed, so callers can construct a working model
without wiring a backbone by hand. The public API is preserved:
`pyvisim.neural_networks.SiameseNeuralNetwork` still resolves to the same class
and all existing constructor calls remain valid.
